pyproject.toml
[tool.poetry]
name = "book-managment"
version = "0.1.0"
description = ""
authors = ["SofiyaYehorova <yehorova66@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
djangorestframework = "^3.15.1"
mysqlclient = "^2.2.4"
drf-yasg = "^1.21.7"



[tool.poetry.group.dev.dependencies]
isort = "^5.13.2"
django-dotenv = "^1.4.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
